requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1666764652625/work
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1606236101993/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1678108872112/work
chex==0.1.7
contourpy==1.1.0
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography-split_1672672380968/work
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
dm-tree==0.1.8
dominate==2.7.0
etils==1.3.0
executing==1.2.0
filelock==3.12.0
flax==0.7.1
fonttools==4.40.0
fsspec==2023.5.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
grpcio==1.54.0
huggingface-hub==0.14.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipdb==0.13.13
ipython==8.12.2
jax==0.4.13
jaxlib==0.4.13
jedi==0.18.2
jimm==0.0.1
jittor==1.3.7.15
kiwisolver==1.4.4
lazy_loader==0.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-dtypes==0.2.0
msgpack==1.0.5
nest-asyncio==1.5.7
networkx==3.1
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.2.3
packaging==23.1
parso==0.8.3
Pbar==1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1675487152289/work
prompt-toolkit==3.0.38
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments==2.15.1
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1680037383858/work
pyparsing==3.1.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1682535435083/work
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.20.0
scipy==1.9.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data==0.6.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.6
tensorstore==0.1.40
tifffile==2023.4.12
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
traitlets==5.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
wcwidth==0.2.6
Werkzeug==2.3.4
zipp==3.15.0
